Move to core and document.
